// api/src/__generated__/products.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: xN1/EbNrE0DLY6b3dMGT0Ru7Ip9qWhGFJx9llLm/XhHgxqldbrlCSEJTctfPFNFg4VMwm8e4R0+YikB9kFiUxg==
 */

// eslint:disable
// tslint:disable

import Categories from './categories'

interface Products {
  /**
   * @default true
   */
  active: boolean
  /**
   * @default 0
   */
  adjustment: number
  category_id: Categories['id']
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at: (Date) | null
  /**
   * @default 0
   */
  discount: number
  /**
   * @default nextval('products_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'products_id'}
  name: string
  /**
   * @default 1
   */
  package_size: number
  /**
   * @default false
   */
  pant: boolean
  price: number
  purchase_price: number
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at: (Date) | null
}
export default Products;

interface Products_InsertParameters {
  /**
   * @default true
   */
  active?: boolean
  /**
   * @default 0
   */
  adjustment?: number
  category_id: Categories['id']
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at?: (Date) | null
  /**
   * @default 0
   */
  discount?: number
  /**
   * @default nextval('products_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'products_id'}
  name: string
  /**
   * @default 1
   */
  package_size?: number
  /**
   * @default false
   */
  pant?: boolean
  price: number
  purchase_price: number
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at?: (Date) | null
}
export type {Products_InsertParameters}